feat: header commponent for reuse in every section
