+++ src/components/index.js
@@ -1,2 +1,4 @@
+export { default as AboutCommunity } from "./aboutCommunity";
+export { default as header } from "./header";
 export { default as Hello } from "./Hello";
 export { default as stats } from "./stats";
